# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LifeLoop API",
    "version": "1.0.0",
    "description": "The backend API powering LifeLoop — scanning, AI recommendations, community, gamification, and notifications. All endpoints except registration/login require a valid session cookie (obtained by logging in)."
  },
  "servers": [{ "url": "/api", "description": "Local server" }],
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "lifeloop_token" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "INVALID_NAME" },
          "message": { "type": "string", "example": "Please tell us what the item is." }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "isAdmin": { "type": "boolean" }
        }
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Create an account",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["name", "email", "password"], "properties": {
            "name": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string", "minLength": 6 }, "rememberMe": { "type": "boolean" }
          } } } }
        },
        "responses": {
          "201": { "description": "Account created", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email already registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Sign in",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "properties": {
            "email": { "type": "string" }, "password": { "type": "string" }, "rememberMe": { "type": "boolean" }
          } } } }
        },
        "responses": {
          "200": { "description": "Signed in", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
          "401": { "description": "Incorrect credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "summary": "Get the currently signed-in user",
        "responses": {
          "200": { "description": "Current user", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
          "401": { "description": "Not signed in", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}